Add schemas.dhall for use with dhall rewrite tool
The `dhall` command has a `rewrite-with-schemas` subcommand, which can
be used to clean up dhall that has been generated by the likes of
`yaml-to-dhall`.

This `dhall rewrite-with schemas` command requires a record of
schemas, where "schema" is of the form of the kind of record used for
record completion; i.e.: `{ Type: Type, default: {}}`

There are some of the types from this repo that are missing from this
schema file, because they're not currently in the appropriate
form. Those missing are:

- APIServerConfig
- AlertmanagerConfig
- AlertmanagerConfigList
- AlertmanagerConfigSpec
- Endpoint
- NamespaceSelector
- Route
- SecretOrConfigMap
- StorageSpec